requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Jinja2==3.1.4
beautifulsoup4==4.12.3
html5lib==1.1
mysql-connector-python==9.0.0
email-validator==2.2.0